Fix Boost Mobile and Sprint 3G
Removed bearer for Boost Mobile and Sprint eHRPD and LTE APNs.
The eHRPD APN was set to bearer 13 even though it is valid for
all 1x/EVDO/eHRPD connections. This was causing no APN to be
selected when an EVDO connection was established and radioTech
!= 13 (eg EVDO_REV_A = 8). The LTE bearer must also be unspecified
so telephony does not then select the eHRPD APN since any unspecified
bearer is potentially valid for both 3G and LTE. Regardless, the
bearer does not need to be specified since the system selects the
correct APN for 3G or LTE automatically.

Change-Id: I0593a9bd24fac6e8e0872be2069e1bc861fe9df1